hugo/server: Rename lib/hugo.js as lib/nominate.js
